IS-1 The starter now exposes a MetadataResolver bean
