+++ pom.xml
@@ -6,7 +6,7 @@
   <groupId>se.swedenconnect.spring.saml.idp</groupId>
   <artifactId>spring-saml-idp-parent</artifactId>
   <packaging>pom</packaging>
-  <version>1.0.0</version>
+  <version>1.0.1-SNAPSHOT</version>
 
   <name>Sweden Connect :: Spring SAML Identity Provider :: Parent POM</name>
   <description>Parent POM for Spring SAML Identity Provider libraries</description>
@@ -44,9 +44,9 @@
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <java.version>17</java.version>
 
-    <spring.boot.version>2.7.11</spring.boot.version>
-    <spring.version>5.3.27</spring.version>
-    <bc.version>1.73</bc.version>
+    <spring.boot.version>2.7.14</spring.boot.version>
+    <spring.version>5.3.29</spring.version>
+    <bc.version>1.76</bc.version>
   </properties>
 
   <repositories>
