+++ saml-identity-provider/src/main/java/se/swedenconnect/spring/saml/idp/Saml2IdentityProviderVersion.java
@@ -24,7 +24,7 @@ public final class Saml2IdentityProviderVersion {
 
   private static final int MAJOR = 1;
   private static final int MINOR = 0;
-  private static final int PATCH = 0;
+  private static final int PATCH = 1;
 
   /**
    * Global serialization value for SAML Identity Provider classes.
