+++ starter/pom.xml
@@ -10,7 +10,7 @@
   <parent>
     <groupId>se.swedenconnect.spring.saml.idp</groupId>
     <artifactId>spring-saml-idp-parent</artifactId>
-    <version>1.0.0</version>
+    <version>1.0.1-SNAPSHOT</version>
   </parent>
   
   <name>Sweden Connect :: Spring SAML Identity Provider :: Spring Boot Starter</name>
